Faster response when receiving hex
QTextDocument is based on blocks, which represent a paragraph(new line).
If a single line is too long, Qt will takes a lot of time to insert something into it.
The time complexity for appending everything to a single line seems to be equal to or greater than O(n),
which means that, the longer the orignal line is, the more time it will take to append even a single char on it.
If I set a length threshold of every line, the response speed will be fine.
There should be a REAL QPlainTextEdit which use a "QPlainTextDocument" or even a QString as the content holder.
The performance of a most common widget should be considered.