Prevent crash on MSVC 32-bit build
The WinRT API might show `80040154 Class not registered` when running
`getRadioFromAdapterId()` and returns a null Radio object. I guess it's
caused by the platform configuration of the WinRT API set to `Any CPU`,
but I'm running the 32-bit build on a 64-bit machine. In this case, the
Bluetooth is disabled. However, I don't know if this build is working on
a native 32-bit machine.